Moved utils files about workspace in a new folder
Easier to work in files like this
