+++ modules/blockly/utils/workspace_management/manage_slash_commands.js
