+++ modules/blockly/utils/workspace_management/validate_workspace_functions.js
